change: MapResourceLoader移动到org.beetl.core.resource
