+++ beetl-core/src/test/java/org/beetl/ext/nutz/BeetlViewMakerTest.java
@@ -6,8 +6,6 @@ import static org.mockito.Mockito.when;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.Reader;
-import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -18,10 +16,7 @@ import javax.servlet.ServletOutputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.beetl.core.GroupTemplate;
-import org.beetl.core.Resource;
-import org.beetl.core.ResourceLoader;
-import org.beetl.core.exception.BeetlException;
+import org.beetl.core.resource.MapResourceLoader;
 import org.nutz.mvc.View;
 
 import junit.framework.TestCase;
@@ -86,38 +81,4 @@ public class BeetlViewMakerTest extends TestCase {
         when(req.getAttribute("obj")).thenReturn(re);
         return req;
     }
-}
-
-class MapResourceLoader extends HashMap<String, String> implements ResourceLoader {
-    
-    private static final long serialVersionUID = 5361058943317379270L;
-    
-    public boolean isModified(Resource key) {
-        return true;
-    }
-    public void init(GroupTemplate gt) {}
-    public String getResourceId(Resource resource, String key) {
-        return key;
-    }
-    public Resource getResource(final String key) {
-        return new Resource(key, this) {
-            public Reader openReader() {
-                String val = get(key);
-                if (val == null) {
-                    BeetlException ex = new BeetlException(BeetlException.TEMPLATE_LOAD_ERROR);
-                    ex.resourceId = key;
-                    throw ex;
-                }
-                return new StringReader(val);
-            }
-            
-            public boolean isModified() {
-                return true;
-            }
-        };
-    }
-    public boolean exist(String key) {
-        return containsKey(key);
-    }
-    public void close() {}
 }
